No more Bug of flashing wrong tiles (uv) !!
Took time to find this simple bug!
The problem was in the asynchronous load of multi textures for the same
tiles. the tile was visible when the length of the texture was equal to
the nb of textures needed but it's wrong as it's possible that
subTexture n comes before subTextures n-1 and in Javascript :  array []
//-> size: 0; array [10]="wao" //-> size: 11