Setup end-of-line normalization
In the repository, all text files will have LF end of lines.
When checked out, in the working directory, they'll use whatever EOL
is configured for your Git installation (core.eol, which defaults to
the platforms native line ending, or CRLF if core.autocrlf is set)

See the gitattributes manpage for details on line ending normalization.