Add earlyDropFunction attribute to Commands
Allow the user to specify a function to decide if a Command
must be cancelled.

Note that children creation commands are tagged as non-cancellable.
That's because children creation only happens when a node has no children.

If 1 of the child creation request is cancelled the node
will be in an invalid state: 3 children and no way to create
the last one.

Fix #82 'Tiles are not displayed'