Refactor source layers and WMTS_Provider
With this change, (data source) layers are only known by the Quadtree.
They are used as an argument to Command instead of having the Provider
aware of them.
This allow several simplifications:
  * a single Provider (e.g WMTS_Provider) instance can be used, since data (layer)
and processing (network fetch, parsing) are now decoupled
  * downsampling logic can be moved to NodeProcess (for now) instead of living in a
Provider. This will help reusing the same logic with other data source.