Fix suboptimal KMZ cache in KML_Provider
The cache now stores promises, whether resolved or not. That way, when
the parseKML method is called a second time for the same KMZ, before
that KMZ's loading has resolved, it will simply return the cached promise
instead of loading the file again. Once resolved, the promise can still
be used directly, so it's kept in the cache (no need to replace it).