Fix concurrency bug in ProjectiveTexturingMaterial.init()
In case init() is called a second time before the it has resolved,
it now returns the same (unresolved) promise instead of a resolved
promise of 'null' (because _shaderMat wouldn't have been initialized
yet). Once resolved, the promise can still be used directly, so it's
only a matter of creating one only once.