Limit command cancelation to refinement commands
The current implementation of BrowseTree and NodeProcess create a dead end if
initial imagery/elevation textures are canceled.

NodeProcess keeps the parent displayed until the children are ready, and
BrowseTree doesn't call NodeProcess.SSE on the children because the parent
is displayed.
So if one of the children's initial command is canceled, NodeProcess will never
issue new commands because it'll never process these children nodes and the
parent will be stuck forever with childrenLoaded() returning false.

Limit NodeProcess.SSE complexity