Fix race condition in the initialisation of the Scene.
Doing Quadtree.init with the geometry layer triggers and update of the
scene, which later causes subdivision to be made to the first 3 levels.
When this subdivision is made, we want to be absolutely sure that
LayerConfiguration is complete. Otherwise, layers textures won't be
downloaded if they are added after this initialisation.

FYI, this race conditions was triggered (most of the time) in Firefox but not in chromium.