fix: increase priority of non-loaded base nodes
Fix a bug very visible on initial load: sometimes the left (resp. right)
part of the globe will refine before the right (resp. left) is even loaded.

So we end up with a detailled view on one side, and a big empty area on the
another one.

This changes prevents this, by giving an increased priority to low-level nodes.