Chore: Use babel to inline glsl file using webpack loader
As we need to inline glsl files for packaging, but *without* bundling
them like webpack does, we just deal with glsl inlining with babel
instead of webpack when building our lib folder. We need a way to tell
babel where and how to process these glsl though, and that's where
babel-plugin-webpack-loaders comes into play.

We still let webpack deal with glsl files when building our bundles, or
when using webpack-dev-server, hence the custom babel config in
webpack.config.js