feature: allow geometry layer to implement partial update
Before this commit there was no link between an update starting and the cause
of this update being fired.

Now the source of notifyChange() calls are stored, and the layer.preUpdate()
method receive them as an argument.
This way preUpdate can choose a better starting point from the update.

As an example I've modified GlobeView/PlanarView to use this feature: updates
now start at the common ancestor (if any) of all the notifyChanges() sources.