refactor: store all node's textures in texture-atlas
Texture atlas usually perform better than individual textures.
It also allows to reduce the number of Sampler2D to the number of layer,
instead of the number of textures.

I didn't merged all the layers in a single atlas, because updating this
huge texture is slow (at least using the naive way of building a THREE.Texture
from an atlas. It may be feasible to access directly WebGL functions like
glTexSubImage2D).

Currently the texture packer is naive and probably only works for our test
data, but it should be easy to implement a more robust solution.
We could also consider modifying the AtlasBuilder to fit more textures in the
same atlas - not necessarly textures from the same tiles.

Last but not least, without this commit iTowns crashes on my phone :)