feature: add DEMUtils module
This module has helpers methods to read the elevation used by the displayed tiles
at some specific coords.

It doesn't return the absolute correct value but instead the displayed one, which
is what we really want most of the time to be able to position objects correctly.

It supports both float texture (xbil) and color textures.