feature: add new FirstPersonControls
This control is very similar to FlyControls but the major difference is that
in this control the camera cannot roll.
This is handy in app using a planar mode where the roll movement is never wanted.

This control can also be used to create a PanoramaControls, just set moveSpeed to 0.

At some point we'll probably refactor all our controls under a single class exposing
the different behavior.