fix: make sure camera's matrix is only update by iTowns
We must make sure that camera.matrixWorld is not modified in between mainloop.update
and mainloop.render otherwise we'll have visual inconsistencies.

e.g: there were holes in the planar view when rotating quickly the camera because of
this issue (three.js calls camera.updateWorldMatrix at the beginning of its rendering
code).