feat (3dtiles): use different subdiv. logic for add and sub refinement
This commit changes the way additive refinement is handled for 3dtiles.
Until now we had a single option: when a tile reached its sse threshold,
we would download and display all of its children.

In the additive refinement we can do better and only download the children
that will be visible. This reduces network/mem usage, for the same visual
result.