feat (core): only use threeJsLayer feature when necessary
Using three.js' layers visibility has a big drawback: we're limited to 31 visibility
groups.
It's quite a low amount, so this commit changes the approach used for visibility:
  - GeometryLayer uses 'layer.object3d.visible' instead
  - layers attached to GeomLayer (like the TileDebug one) still use threeJsLayer,
    because their elements are attached to another layer's objects so they don't
    all share a common parent

Another change included is that now it doesn't make sense to use view.scene as a
default object3d parameter, so this commit switches to a Group instead.