chore(packaging): do not expose peer dependencies when using the modules
When people uses a bundler, they can directly import these no need to
pollute our namespace with itowns.THREE, proj4 etc.

However, for people using our bundle, we still expose those
peerDependencies, are they are included in it anyway. In the future, we
might provide 2 bundles: one full bundle like it is today, and a bundle
relying on ambiant peerDependencies (attached to the windows object).