feat (controls): add collision camera/tiles's obb in globeControls
add collision between camera's bounding and tiles's oriented bounding box. Depending on the distance of the camera with obbs, we add a slowdown or constrain to the movement. this constraint or deceleration is suitable for two types of movement MOVE_GLOBE and ORBIT. These constraint or deceleration are inversely proportional to the camera/obb distance