refactor(core): use common geometry in meshTile
 * identical geometries are reused to build the tileMeshs
 * the bounding boxes is calculated by THREE.js
 * bounding boxes can use the Parent's worldMatrix (tile).

BREAKING CHANGE
The geodesic normal in the center of the tile's geometry is aligned with the mesh's local Z axis.
Previously, the tiles were not oriented. their local landmark had the same orientation as the world landmark. You must take into account the orientation of the tile if you add objects to this tile.

Fixes:
 * Ellipsoid tile: The tile's geometry normal is computed with correct geodesic normal
 * reduces obb calculation errors see #87 issues
 * errors obb/culling for pano tile

Performance
 * the processes time (to build tiles in TileProvider.executeCommand) and used memory (geometryTile) are divided by ~3