refactor(core): unify mesh picking methods
The idea of this PR is to expose a single entry point for querying the view.

This method belongs to the View.

Here are a few example usage:
  * view.pickObjectsAt({ x, y })
  * view.pickObjectsAt({ x, y }, 'wfsBuilding')
  * view.pickObjectsAt({ x, y }, 'wfsBuilding', myLayer)
  * view.pickObjectsAt({ x, y }, 'wfsBuilding', myObject3d)

It supports either layers, layer's id or THREE.Object3D.
By default it'll use THREE.Raycaster, but some layer can provide their own implementation.

Currently the layer with a custom implementation are:
  - layers using TileMesh (globe, planar, panorama)
  - pointclouds