feat(core): compute TMS coordinates from top to bottom
As it can be read in the link at the end, TMS coordinates differ from
other "TMS" coordinates from Google Maps, OSM, Bing etc... The first one
goes from bottom to top, while the latter top to bottom. In this
commit, the parameter bottomLeft in the method computeTMSCoordinates
can be set to true to obtain the same result as Google Maps and others.

Source:
https://alastaira.wordpress.com/2011/07/06/converting-tms-tile-coordinates-to-googlebingosm-tile-coordinates/
https://wiki.osgeo.org/wiki/Tile_Map_Service_Specification
https://www.planet.com/docs/reference/tile-services