refactor(core): Homogenize all providers as static
There is no reason to keep an instance of each provider, and moving them
to static like some other one (PointCloudProvider for example) do no
harm, but homogenize all providers to one structure.

Also note that there are two required functions in a provider:
preprocessDataLayer and executeCommand, that are verified when adding a
provider, in the scheduler.

BREAKING CHANGE: KML_Provider and BuildingBox_Provider have been
removed, as well as getPointOrder from WFS_Provider.