refactor(core): change view.notifyChange semantics
This commit changes the semantics of view.notifyChange function to something more
useful:
  changeSource now drives layer update.

This implies that no changeSource -> no update.

'changeSource' can be either a layer, an object from a layer or the camera.

The signature also changed slightly:
   notifyChange(needsRedraw = false, changeSource = undefined)
becomes:
   notifyChange(changeSource = undefined, needsRedraw = true)

Some examples usage:
  - notifyChange(): the scene will be redraw, but no update will be performed. This
is useful if a user adds some objetcs to the scene and wants to request a redraw but
do no need a full update of all layer (eg: a measure tool drawn with threejs lines)
  - notifyChange(XXX): only the layer XXX (or the layer to which XXX belongs) will
be updated. So if you have 100 layers, and layer #4 has a new texture, it will not
cause a full update of other layers.
  - notifyChange(camera): this is equivalent to the previous notifyChange(true): all
layers will be updated.

MainLoop also handles changeSource filtering and could later even handle ancestor
lookup (if we build a common ancestor lookup interface).

BREAKING CHANGE: change view.notifyChange signature and semantics