refactor: align Feature2Mesh and Feature2Texture
The two modules are very similar: they take a feature or an array of
feature as arguments, and return either a mesh or a texture. It makes
sense to regroup them under the same scheme.

They have been moved to another folder, src/Transform. They are not
available anymore under Feature2*.convert, but rather directly calling
feature2Mesh() or feature2Texture().

The documentation was completed in the same time.

BREAKING CHANGE: Feature2Mesh.convert doesn't exist anymore, in favor of
feature2Mesh, and the same applies for
Feature2Texture.createTextureFromFeature in favor of feature2Texture.