refactor: change the geojson parsing of multi-
When parsing a multi polygon, line or point, the result was a modified
Array with custom properties, not respecting the original object. It has
been replaced by an Object similar to the definition of a Feature
Collection.

In the same time, Feature2Mesh and Feature2Texture have been slightly
rewritten to match the new structure. The rewrite has added a fix in the
extrusion of features, and simplified the processing in both files.