feat(core): add a Parser registry
Following #723 and the model of Providers, a Parser registry has been
partially added to the Scheduler. You can now register a Parser using
scheduler.addFormatParser(formatName, parser).

In the Scheduler queue, the Parser will be called right after the
Provider. It takes the datablob returned by the Provider, parses it, and
return it.

The parser selection relies on the format setted in the layer. For
layers that don't have a format, or have a format that is not supported,
a fake parser, returning immediatly the blob, is setted.

This commit introduces three parsers: GeoJson, Gpx and Kml. This allows
to refactor a bit the RasterProvider, in hope of letting go of it in the
future.

BREAKING CHANGE: GpxParser doesn't return a THREE.Mesh anymore.