fix(3dtiles): dont stop subdivision at tileset boundaries
In this scenario:
  - a tileset references an external tileset at node N
  - the subdivision algorithm decides to stop at N and display N
We must go one step further and display the content of the root
node R from the external tileset.

The reason is that R and N are the same node regarding content,
but are stored in different tilesets (for technical reasons)  and
thus considered as different node in iTowns (with N having R as
its only child).