feat(pointclouds): make point picking independant of the source type
This commit reworks the picking code for pointclouds to allow pointclouds
from 3dtiles to be picked in the same way.

The default picking behavior for a 3dtiles layer is still the raycasting
because a 3dtiles layer can be composed of several objects type.
But this commit allows to use 'Picking.pickPointsAt' if one wants to select
points (see test/examples/3dtiles.js).