fix(parser): change the internal structure of geojson
There was a problem with the destructuration when parsing multitypes in
the GeojsonParser, so the internal used structure has been changed to
fix this issue.

A test has been added at the same time, verifying part of this issue as
it could happened when the features were out of the specified filtering
extent.