test: reduce test time for examples
The previous way of testing the examples was relatively slow, as a page
was created for EACH test, even for a same group of test.

Now, there is only one page created for all tests, that is only reloaded
when changing the test example. This reduces time spent on this tests
drastically.