refactor(gis): change ellipsoidSizes from function to object
The ellipsoidSizes export in Coordinates was a function, which was
creating an object each time it was called. It has now be changed to an
object, meaning that we don't have to execute the function each time.

BREAKING CHANGE: ellipsoidSizes from Coordinates.js is no longer a
function but an object.