refactor(shader): manage chunk directly into shaders
Before, shader chunks were added manually in JS code for each material.
Now they are directly added in shaders. To use it, ShaderChunk has been
added with the method install, that is call in LayeredMaterial.

In the same time, the colorAtIdUv unrolling has been moved to the
shader, and unrolled using the new ShaderUtils.unrollLoop method and the
pragma unroll_loop.