refacto(source): deprecate adding a source with only a configuration
Documentation has been rewritten to reflect change. New properties have
been added (see breaking changes).
The only one untouched source is StaticSource, as it is meant to be
removed in the future, and is quite unstable to touch.

BREAKING CHANGE: the origin property of TMSSource has been replaced by a
boolean, isInverted, that serves the same goal, but using a boolean.
BREAKING CHANGE: the protocol property no longer exists, instead to test
the type of Source, one needs to use the property isSource, adapted to
each type: for example, isWMTSSource tells if the source is WMTSSource.