fix(core): implement preSSE once and fix the formula
3dTilesProcessing and GlobeTileProcessing use the same wrong SSE formula.

PR #633 and #818 proposed to fix this, but wasn't merged.

This commit fixes the computation, and comment it so the details should
be clearer and less dependant on an external reference that might go away.

The fixed bug can be seen easily in the oriented image example: if you
zoom in, the moutains behind the photos loses precision instead of increase it.

This should be fixed, and I added a unit test for this problem.

BREAKING CHANGE: Computing screen space error is changed, so tile subdivision are different.