refactor(layer): replace all layer.type by layer.isLayer
In 36c6d03 (PR #951), the isXXXLayer has been added, but wasn't used
everywhere. The change has been reflected in this commit to remove all
occurence of layer.type, leaving only the use for the debug layer (to be
removed).

DEPRECATION NOTICE: type is no longer a property of a Layer