feat(example): add a DragNDrop module in examples/
This module can be added to a web page using the `<script>` tag or used
as a module in a built webapp. The user can then bind the viewer to it,
and register some file extensions to support. It is very simple, but can
be used for quick examples where people want to test files without
having to write code.