feat(tile): add a refreshMaterial method
In a case where you have a server, with the data updated from what is
displayed, you want to update it without destroying the whole layer and
recreating it. If you know the tile you want to update, you can use the
new `TileMesh#refreshMaterial` method.

In the same time, a new `FORCE` state has been added to
`LayerUpdateState`, that is used when doing this refresh.