refacto(cache): ditch Map() in Cache for simple tree (#1264)
refacto(cache): ditch Map() in Cache for simple tree

The previous Cache was working on a single Map, which can become really
slow and big after a few minutes of navigation in iTowns.

Now, the Cache is composed of a single tree, that can contain node up
to 3 levels. It simplifies things attached to a layer, a source or
others things like a zoom value. Instead of having a complex string
composed with all those information, it now uses several key accross all
the node of the Cache.

In term of performance, the proposed implementation is 20 times faster
than the previous one.