fix(core): remove objects of a GeometryLayer from the Scene
The property object3d (an instance of THREE.Object3D, most of the time
THREE.Group) was not correctly removed for a GeometryLayer, when calling
View#removeLayer.

Now, if the layer is a geometry one, the GeometryLayer#delete method is
called once more, and should remove the object3d property from its
parent (usually a THREE.Scene).