fix(feature): return style when filtering features
In FeatureUtils, when filtering features under a coordinate, the style
of each feature is returned, in order to be used (in FeatureToolTip for
this case). It is a fix that corrects f703723.

This is not the best fix to do, but the quickest at the time. In the
future, I recommend getting rid of the FeatureUtils and do all the work
directly in Feature/FeatureCollection, and return a partial Feature or
FeatureCollection depending on the case.