fix(view): use correct DOM element and offset
Before that, when the viewer div was not in fullscreen, or covering the
whole page (a case we do not test a lot), there was a bad conversion
between a screen coordinates and the viewer div coordinates. This was
due to a change in e3ff1591 and the use of `clientXY` instead of
`offsetXY`.

This change comes with a tiny test unit change when moving the mouse in
`globecontrols`, but we may need more deep testing here. I won't do it
now to have this fix in the 2.19.0 release.

Fixes #1333