feat(renderer): emit a (in)visible event for LayeredMaterial
Knowing the visibility of a `LayeredMaterial` is useful. For example in a
`GlobeView`, if you zoom in, "parent" tiles seems hidden; in fact, there
are not, it is only their material (so `LayeredMaterial`) that is set to
not visible.

Adding an event when changing this property can be useful to hide others
things, like in `TileDebug`, or in later PR to come (#1303 for example).