fix(demutils): interpolate values considering noDataValue
When interpolating elevation values, in case of a precise search, if one
of the point has the value of `noDataValue`, it can cause errors if
`noDataValue` has a great value.

This commit ensures that this is taken into account, by replacing those
values by undefined, and thus still giving the appropriate result at the
end.