feat(style): add partial support for steps in vector tiles
If there are `stops` in any property of the Style, they will be taken
into account to override `minzoom` and `maxzoom` to reflect those
`stops`.

Note that this is a partial support, as the Mapbox Style specs have evolved:
https://docs.mapbox.com/mapbox-gl-js/style-spec/expressions

This commit is also moving a lot of things that were in
`VectorTileParser` back to `VectorTilesSource`, to keep things more
independant.