refacto(style): get anchor position values instead of CSS
Before, the `Style#getTextAnchorPositionInCSS()` method was returning a
string, to be used directly in CSS.

Now, it has been modified to `Style#getTextAnchorPosition()`, and
returns an array containing the `x` and `y` proportions to applied to a
`Label` for its anchor.

BREAKING CHANGE: `Style#getTextAnchorPositionInCSS` has been renamed to
`Style#getTextAnchorPosition`.