fix(source): set the Style.minzoom correctly in VectorTilesSource
The previous behavior was as follow:
- check the presence of `minzoom` for a specific layer
- if it is not present, use `2` has a default value
- insert this value to the `stops` array
The resulting problem is that if `minzoom` is not defined and the first
`stop` value was greater than 2, it was still adding a `2 to first stop`
step.

A solution has been added here, to consider that the `minzoom` is either
the specified minzoom, or whichever is the highest between 2 and the
littlest `stop`.