example: add a vector tile example with style drag and drop
This new example allows to drag and drop a valid style file, with the
sources of the stream specified in it, in order to test it. It is a
relatively simple tester, as nothing too complex has been tested with
it.

The DragNDrop is not used here, because it is only adapted for files
sources, not streams.