feat(style): add Style inheritance
It is now possible to specify a `Style` parent for another `Style`. That
way, it is possible to define a global `Style` for a `ColorLayer`, while
defining one with property specific to a `FeatureGeometry` for example.
The `Style` of the `FeatureGeometry` will then only have the minimum
requirement, and it will read missing properties from its parent. It is
a dynamic reading, so it the parent values are changing, it is changing
as well for the child style.

This is a proposition for improvement in #1318.

BREAKING CHANGE: `Style#text.halo` is not available anymore, and each
property of it as been place in the `text` object, with the `halo`
prefix: `haloWidth`, `haloColor` and `haloBlur`.