fix(label): correctly hide dom
There was a problem for a specific situation. Let's consider that there
is a level 2 node, with four level 3 nodes. The L2 node contains no
DOM elements, while there are some in the L3 nodes. So the DOM in L3
have no parents, they are the first in the hierarchy.  So if we find L2
hidden, we don't go further, and thus inside L3 nodes, and don't hide
the DOM.

This commit fixes this behaviour, by allowing to go further in the
hierarchy until DOM elements are found.